third_party/blink/web_tests/external/wpt/html/canvas/element/drawing-text-to-the-canvas/2d.text.drawing.style.nonfinite.spacing.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.text.drawing.style.nonfinite.spacing</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.text.drawing.style.nonfinite.spacing</h1>
<p class="desc">Testing letter spacing and word spacing with nonfinite inputs</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Testing letter spacing and word spacing with nonfinite inputs");
_addTest(function(canvas, ctx) {

_assertSame(ctx.textLetterSpacing, 0, "ctx.textLetterSpacing", "0");
_assertSame(ctx.textWordSpacing, 0, "ctx.textWordSpacing", "0");

function test_word_spacing(value) {
  ctx.textWordSpacing = value;
  ctx.textLetterSpacing = value;
  _assertSame(ctx.textWordSpacing, 0, "ctx.textWordSpacing", "0");
  _assertSame(ctx.textLetterSpacing, 0, "ctx.textLetterSpacing", "0");
}
test_word_spacing(NaN);
test_word_spacing(Infinity);
test_word_spacing(-Infinity);


});
</script>
